fix(meet-ext): move test-only handlers out of content-script entrypoint
Chrome loads MV3 content scripts as classic scripts, so a top-level
`export` token in `dist/content.js` can make the whole bundle fail
to parse at load time and silently kill every Meet handler in
production. The prior PR added `__handleSendChat` / `__handleCameraToggle`
/ `__enqueueSendChat` exports to `content.ts` as test hooks and
tripped this hazard.

Move the send-chat and camera-toggle handlers (plus the per-tab
serialization queue) into a new `handle-send-chat.ts` helper module,
import it from `content.ts` as a named-import side effect, and drop
all top-level `export` tokens from `content.ts`. Tests import the
handlers from the helper module directly.

Verified `dist/content.js` now contains zero `export` tokens after
`bun run build`.
