+++ skills/meet-join/meet-controller-ext/src/content.ts
@@ -24,18 +24,16 @@
  * in that case because the scrapers require an admitted meeting.
  */
 import type {
-  BotCameraDisableCommand,
-  BotCameraEnableCommand,
-  BotSendChatCommand,
   BotToExtensionMessage,
-  ExtensionCameraResultMessage,
-  ExtensionSendChatResultMessage,
   ExtensionToBotMessage,
 } from "../../contracts/native-messaging.js";
 import { BotToExtensionMessageSchema } from "../../contracts/native-messaging.js";
 
-import { disableCamera, enableCamera } from "./features/camera.js";
-import { type ChatReader, sendChat, startChatReader } from "./features/chat.js";
+import {
+  enqueueSendChat,
+  handleCameraToggle,
+} from "./handle-send-chat.js";
+import { type ChatReader, startChatReader } from "./features/chat.js";
 import { runJoinFlow } from "./features/join.js";
 import {
   startParticipantScraper,
@@ -193,37 +191,6 @@ function startMeetingSession(
  */
 let activeSession: MeetingSessionHandle | null = null;
 
-/**
- * Per-tab serialization chain for `send_chat` handling. `sendChat` mutates
- * a single shared textarea (`.value = text`) and then clicks the send
- * button, so overlapping requests would otherwise race on the composer.
- * Chaining onto this promise forces strict arrival-order processing while
- * leaving the `onMessage` listener synchronous (the listener returns
- * immediately; handling happens off-thread).
- */
-let sendChatQueue: Promise<void> = Promise.resolve();
-
-/**
- * Chain a `send_chat` invocation onto the per-tab queue so it runs
- * strictly after any prior in-flight `sendChat` call has completed.
- * Extracted from the inline listener wiring so tests can drive the queue
- * directly (Bun caches ESM modules across tests, so the listener's
- * `chrome.runtime.onMessage.addListener` registration happens once at
- * first-import time — not re-runnable against a fresh fake chrome on
- * each test).
- */
-function enqueueSendChat(cmd: BotSendChatCommand): Promise<void> {
-  sendChatQueue = sendChatQueue
-    .catch(() => {
-      // A prior `handleSendChat` rejection must not poison subsequent
-      // sends — the handler catches its own errors and reports them via
-      // `send_chat_result(ok=false)`, so any rejection here is a bug we
-      // still want to isolate from the next request.
-    })
-    .then(() => handleSendChat(cmd));
-  return sendChatQueue;
-}
-
 chrome.runtime.onMessage.addListener(
   (
     raw: unknown,
@@ -357,132 +324,3 @@ async function handleJoin(
   }
 }
 
-/**
- * Execute a {@link BotSendChatCommand} and emit a matching
- * {@link ExtensionSendChatResultMessage} back to the background. Errors
- * are caught and surfaced via `ok: false` so the bot can correlate the
- * failure with the originating request.
- *
- * Threads an `onEvent` sink + `window` reference through to
- * {@link sendChat} so the runtime `meet_send_chat` tool path emits
- * `trusted_type` (for the composer) and `trusted_click` (for the send
- * button) just like the consent-post path does inside `runJoinFlow`.
- * Without this, Meet's `isTrusted` gate silently swallows both the
- * synthetic composer input and the JS `.click()` on the send button —
- * every post-admission send would no-op on production Meet builds that
- * enforce the gate.
- */
-async function handleSendChat(cmd: BotSendChatCommand): Promise<void> {
-  const sendToBot = (event: ExtensionToBotMessage): void => {
-    try {
-      void chrome.runtime.sendMessage(event);
-    } catch (err) {
-      console.warn("[meet-ext] sendMessage failed:", err);
-    }
-  };
-
-  let reply: ExtensionSendChatResultMessage;
-  try {
-    await sendChat(cmd.text, {
-      onEvent: sendToBot,
-      // Pass the live `window` so `sendChat` can compute screen-space
-      // coordinates for the send button's `trusted_click`. Mirrors the
-      // fallback that `postConsentMessage` relies on in `features/join.ts`.
-      window: globalThis as unknown as {
-        screenX: number;
-        screenY: number;
-        outerHeight: number;
-        innerHeight: number;
-      },
-    });
-    reply = {
-      type: "send_chat_result",
-      requestId: cmd.requestId,
-      ok: true,
-    };
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    reply = {
-      type: "send_chat_result",
-      requestId: cmd.requestId,
-      ok: false,
-      error: message,
-    };
-  }
-  try {
-    chrome.runtime.sendMessage(reply);
-  } catch (err) {
-    console.warn("[meet-ext] failed to send send_chat_result:", err);
-  }
-}
-
-/**
- * Execute a {@link BotCameraEnableCommand} / {@link BotCameraDisableCommand}
- * and emit a matching {@link ExtensionCameraResultMessage} back to the
- * background. Mirrors {@link handleSendChat}: forwards a trusted_click via
- * `onEvent` so the bot drives the click through xdotool (Meet's isTrusted
- * gate rejects synthetic clicks on bottom-toolbar controls in general, so
- * we assume the camera toggle is gated too and route through xdotool by
- * default). Errors are surfaced via `ok: false` with a descriptive reason.
- */
-async function handleCameraToggle(
-  cmd: BotCameraEnableCommand | BotCameraDisableCommand,
-): Promise<void> {
-  const sendToBot = (event: ExtensionToBotMessage): void => {
-    try {
-      void chrome.runtime.sendMessage(event);
-    } catch (err) {
-      console.warn("[meet-ext] sendMessage failed:", err);
-    }
-  };
-
-  let reply: ExtensionCameraResultMessage;
-  try {
-    const run = cmd.type === "camera.enable" ? enableCamera : disableCamera;
-    const result = await run({
-      onEvent: sendToBot,
-      // Pass the live `window` so the camera feature can compute screen-
-      // space coordinates for the toggle's `trusted_click`. Mirrors the
-      // fallback that `postConsentMessage` / `sendChat` rely on.
-      window: globalThis as unknown as {
-        screenX: number;
-        screenY: number;
-        outerHeight: number;
-        innerHeight: number;
-      },
-    });
-    reply = {
-      type: "camera_result",
-      requestId: cmd.requestId,
-      ok: true,
-      changed: result.changed,
-    };
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    reply = {
-      type: "camera_result",
-      requestId: cmd.requestId,
-      ok: false,
-      error: message,
-    };
-  }
-  try {
-    chrome.runtime.sendMessage(reply);
-  } catch (err) {
-    console.warn("[meet-ext] failed to send camera_result:", err);
-  }
-}
-
-// Export the send-chat + camera-toggle handlers for unit testing. They are
-// wired into `chrome.runtime.onMessage` above when the script loads; the
-// tests import them directly to drive the tool paths end-to-end without
-// needing to fake the chrome.runtime.onMessage dispatcher. Not part of the
-// extension's public surface — the background SW never imports content.ts.
-export { handleSendChat as __handleSendChat };
-export { handleCameraToggle as __handleCameraToggle };
-// Exported so the per-tab serialization queue can be exercised directly
-// by tests. The `chrome.runtime.onMessage` listener is registered once at
-// module-load time (Bun's ESM cache prevents re-running module top level
-// across tests), so tests that want to assert ordering for overlapping
-// `send_chat` frames drive this helper instead of the raw listener.
-export { enqueueSendChat as __enqueueSendChat };
